+++ packages/runtime/src/lib/constants.ts
@@ -10,7 +10,7 @@ export const MAX_CONCAT_SIZE = 4096;
 export const ALGORAND_MIN_TX_FEE = 1000;
 // https://github.com/algorand/go-algorand/blob/master/config/consensus.go#L659
 export const ALGORAND_ACCOUNT_MIN_BALANCE = 0.1e6; // 0.1 ALGO
-export const MaxTEALVersion = 3;
+export const MaxTEALVersion = 4;
 
 // values taken from: https://developer.algorand.org/docs/features/asc1/stateful/#minimum-balance-requirement-for-a-smart-contract
 // minimum balance costs (in microalgos) for ssc schema
@@ -109,12 +109,17 @@ TxnFields[3] = {
   LocalNumByteSlice: 'nbs'
 };
 
+TxnFields[4] = {
+  ...TxnFields[3]
+};
+
 // transaction fields of type array
 export const TxArrFields: {[key: number]: Set<string>} = {
   1: new Set(),
   2: new Set(['Accounts', 'ApplicationArgs'])
 };
 TxArrFields[3] = new Set([...TxArrFields[2], 'Assets', 'Applications']);
+TxArrFields[4] = TxArrFields[3];
 
 export const TxFieldDefaults: {[key: string]: any} = {
   Sender: zeroAddress,
@@ -235,6 +240,11 @@ GlobalFields[3] = {
   CreatorAddress: null
 };
 
+// global fields supported by tealv4
+GlobalFields[4] = {
+  ...GlobalFields[3]
+};
+
 // creating map for opcodes whose cost is other than 1
 export const OpGasCost: {[key: number]: {[key: string]: number}} = { // version => opcode => cost
   // v1 opcodes cost
@@ -259,3 +269,8 @@ OpGasCost[2] = {
  * All other opcodes have cost 1
  */
 OpGasCost[3] = { ...OpGasCost[2] };
+
+/*
+ * tealv4
+ */
+OpGasCost[4] = { ...OpGasCost[3] };