+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Improvements
+* TEALv4 support in `@algo-builder/runtime`:
+    * Added shared space between contracts (opcodes `gload` and `gloads`)
+    * Dynamic Opcode Cost Evaluation
+
 ### Bug Fixes
 * Fixed `yarn add @algo-builder/web` (was failing because of missing dependency `zod` in packages/web).